requirements.txt
PyQt6
faiss-cpu
numpy
requests
sentence-transformers
pypdf
